+++ .eslintrc.cjs
@@ -5,7 +5,13 @@ module.exports = {
     'plugin:@typescript-eslint/recommended',
     'plugin:@typescript-eslint/recommended-requiring-type-checking',
   ],
-  ignorePatterns: ['dist', 'node_modules', 'test', '*.cjs'],
+  ignorePatterns: [
+    'dist',
+    'node_modules',
+    'test',
+    '*.cjs',
+    'openapi-ts-request.config.ts',
+  ],
   plugins: [],
   parserOptions: {
     ecmaVersion: 'latest',