+++ test/genOpenapi.js
@@ -1,14 +1,7 @@
 const { generateService } = require('../dist/index');
 
 generateService({
-  schemaPath: `${__dirname}/example-files/openapi.json`,
-  serversPath: './apis/js-client',
-  isGenJavaScript: true,
-  isGenReactQuery: true,
-});
-
-generateService({
-  schemaPath: `${__dirname}/example-files/openapi.json`,
-  serversPath: './apis/react-query',
-  isGenReactQuery: true,
+  schemaPath: `${__dirname}/example-files/openapi-desc-enum.json`,
+  serversPath: './apis/openapi-desc-enum',
+  isSupportParseEnumDesc: true,
 });